refactor: move Hero components to home/_components and rename them
