+++ app/home/_components/Hero/HeroContent.tsx
